branca: Add (en|de)code interface in Branca struct
This introduces a cleaner and simpler way of creating
branca tokens without specifing the nonce or timestamp
fields unlike the direct public methods with are still
accessible, but too verbose to use.

Creating a branca token can be done via calling Branca::new::(&key)
where the key is a &[u8]. The 'build' method has been replaced with
encode / decode methods that simplify generating tokens with the
Branca struct initialised.

The direct encode and decode methods for the nonce and key were
both changed to &[u8] and the associated methods in Branca struct
now borrows a self variable rather than taking ownership of it.

Added Ring dependency for generating secure nonces. Whether to
change to the rand crate might still be possible.

Lastly, cleaned up redundant code with rust-clippy.